2009-03-09	Joel Sherrill <joel.sherrill@OARcorp.com>
	PR 1376/filesystem
	* libcsupport/include/rtems/libio.h, libfs/src/dosfs/fat_file.c,
	libfs/src/dosfs/msdos.h, libfs/src/dosfs/msdos_dir.c,
	libfs/src/dosfs/msdos_file.c, libfs/src/imfs/deviceio.c,
	libfs/src/imfs/imfs.h, libfs/src/imfs/imfs_directory.c,
	libfs/src/imfs/memfile.c, libfs/src/nfsclient/src/nfs.c: lseek
	handlers should return off_t.

+++ cpukit/libfs/src/dosfs/fat_file.c
@@ -49,7 +49,7 @@ _hash_search(
     fat_file_fd_t			   **ret
 );
 
-static int
+static off_t
 fat_file_lseek(
     rtems_filesystem_mount_table_entry_t  *mt_entry,
     fat_file_fd_t                         *fat_fd,
@@ -940,7 +940,7 @@ _hash_search(
     return -1;
 }
 
-static int
+static off_t
 fat_file_lseek(
     rtems_filesystem_mount_table_entry_t  *mt_entry,
     fat_file_fd_t                         *fat_fd,
